upgrade to dotnet 6

+++ Core/Lyra.Data/Crypto/ECC/ECDsa.cs
@@ -30,7 +30,7 @@ namespace Neo.Cryptography.ECC
             BigInteger trunc = new BigInteger(message.Reverse().Concat(new byte[1]).ToArray());
             if (n.GetBitLength() < messageBitLength)
             {
-                trunc >>= messageBitLength - n.GetBitLength();
+                trunc >>= messageBitLength - (int)n.GetBitLength();
             }
             return trunc;
         }
@@ -50,7 +50,7 @@ namespace Neo.Cryptography.ECC
                     {
                         do
                         {
-                            k = rng.NextBigInteger(curve.N.GetBitLength());
+                            k = rng.NextBigInteger((int)curve.N.GetBitLength());
                         }
                         while (k.Sign == 0 || k.CompareTo(curve.N) >= 0);
                         ECPoint p = ECPoint.Multiply(curve.G, k);
@@ -71,7 +71,7 @@ namespace Neo.Cryptography.ECC
 
         private static ECPoint SumOfTwoMultiplies(ECPoint P, BigInteger k, ECPoint Q, BigInteger l)
         {
-            int m = Math.Max(k.GetBitLength(), l.GetBitLength());
+            int m = Math.Max((int)k.GetBitLength(), (int)l.GetBitLength());
             ECPoint Z = P + Q;
             ECPoint R = P.Curve.Infinity;
             for (int i = m - 1; i >= 0; --i)
